Eliminated unnecessary object, merged with Activity.
